adds stackMoveMutation
