+++ src/components/project/projectDndProvider.tsx
@@ -1,3 +1,4 @@
+import { Project, Stack, Task, httpClient } from '@/axios';
 import {
   DndContext,
   DragEndEvent,
@@ -5,17 +6,30 @@ import {
   DragOverlay,
   DragStartEvent,
   PointerSensor,
+  closestCorners,
   useSensor,
   useSensors,
 } from '@dnd-kit/core';
-import { ReactNode, useState } from 'react';
-import StackCard from '../stack/stackCard';
-import { Project, Stack, Task, httpClient } from '@/axios';
-import { createPortal } from 'react-dom';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { ReactNode, useState } from 'react';
+import { createPortal } from 'react-dom';
+import StackCard from '../stack/stackCard';
+import TaskCard from '../task/taskCard';
 import { AxiosResponse } from 'axios';
 import { arrayMove } from '@dnd-kit/sortable';
-import TaskCard from '../task/taskCard';
+
+export interface stackMoveDto {
+  projectId: string;
+  stackOrder: string[];
+}
+
+export interface taskMoveDto {
+  taskId: string;
+  fromStackId: string;
+  fromStackTaskOrder: string[];
+  toStackId: string;
+  toStackTaskOrder: string[];
+}
 
 export interface IProjectDndProviderProps {
   children: ReactNode;
@@ -26,76 +40,289 @@ export default function ProjectDndProvider({
   children,
   projectId,
 }: IProjectDndProviderProps) {
-  const [activeStackDataStart, setActiveStackDataStart] = useState<
-    Stack | undefined
-  >(undefined);
-  const [activeTaskDataStart, setActiveTaskDataStart] = useState<
-    Task | undefined
-  >(undefined);
-
-  const [currentStackId, setCurrentStackId] = useState<string | undefined>(
-    undefined,
-  );
-  const [activeTaskId, setActiveTaskId] = useState<string | undefined>(
-    undefined,
-  );
-  const [overTaskId, setOverTaskId] = useState<string | undefined>(undefined);
-
+  const [overlayStackData, setOverlayStackData] = useState<Stack | undefined>();
+  const [overlayTaskData, setOverlayTaskData] = useState<Task | undefined>();
+  const [fromTaskStackId, setFromTaskStackId] = useState<string | undefined>();
+  const [activeTaskStackId, setActiveTaskStackId] = useState<
+    string | undefined
+  >();
   const queryClient = useQueryClient();
 
-  const stackOrderMutation = useMutation({
-    mutationFn: async ({
-      stackOrder,
-      projectId,
-    }: {
-      stackOrder: string[];
-      projectId: string;
-    }) => {
-      await httpClient.patch(`/projects/${projectId}`, { stackOrder });
+  const taskMoveMutation = useMutation({
+    mutationFn: async (data: taskMoveDto) => {
+      await httpClient.post('/dnd/taskMove', data);
     },
-    onMutate: ({ stackOrder, projectId }) => {
-      const prevQueryData = queryClient.getQueryData(['projects', projectId]);
+  });
 
+  const stackMoveMutation = useMutation({
+    mutationFn: async (data: stackMoveDto) => {
+      await httpClient.post('/dnd/stackMove', data);
+    },
+  });
+
+  const onDragStart = ({ active }: DragStartEvent) => {
+    if (active.data.current?.type === 'Stack') {
+      setOverlayStackData(active.data.current.stackData);
+    }
+
+    if (active.data.current?.type === 'Task') {
+      setOverlayTaskData(active.data.current.taskData);
+      setActiveTaskStackId(active.data.current.taskData.stackId);
+      setFromTaskStackId(active.data.current.taskData.stackId);
+    }
+  };
+
+  const onDragOver = ({ active, over }: DragOverEvent) => {
+    if (!over) return;
+    console.log('dragOverActiveId', active.id);
+    console.log('dragOverOverId', over.id);
+
+    if (active.id === over.id) return;
+
+    if (
+      active.data.current?.type === 'Task' &&
+      over.data.current?.type === 'Task'
+    ) {
+      if (activeTaskStackId !== over.data.current.taskData.stackId) {
+        //remove task from stack
+        queryClient.setQueryData(
+          ['stacks', activeTaskStackId],
+          (stackQueryRes: AxiosResponse<Stack> | undefined) => {
+            if (!stackQueryRes) return;
+            const updatedTaskOrder = stackQueryRes.data.taskOrder.filter(
+              (taskId) => taskId !== active.id.toString(),
+            );
+
+            return {
+              ...stackQueryRes,
+              data: { ...stackQueryRes.data, taskOrder: updatedTaskOrder },
+            };
+          },
+        );
+
+        //add task to stack
+        queryClient.setQueryData(
+          ['stacks', over.data.current.taskData.stackId],
+          (stackQueryRes: AxiosResponse<Stack> | undefined) => {
+            console.log(stackQueryRes);
+            if (!stackQueryRes) return;
+            const updatedTaskOrder = [...stackQueryRes.data.taskOrder];
+            updatedTaskOrder.push(active.id.toString());
+            return {
+              ...stackQueryRes,
+              data: { ...stackQueryRes.data, taskOrder: updatedTaskOrder },
+            };
+          },
+        );
+
+        setActiveTaskStackId(over.data.current.taskData.stackId);
+        return;
+      }
+    }
+
+    if (
+      active.data.current?.type === 'Task' &&
+      over.data.current?.type === 'Stack'
+    ) {
+      if (activeTaskStackId !== over.id.toString()) {
+        //Remove from active stack
+        queryClient.setQueryData(
+          ['stacks', activeTaskStackId],
+          (stackQueryRes: AxiosResponse<Stack> | undefined) => {
+            if (!stackQueryRes) return;
+            const updatedTaskOrder = stackQueryRes.data.taskOrder.filter(
+              (taskId) => taskId !== active.id.toString(),
+            );
+            return {
+              ...stackQueryRes,
+              data: { ...stackQueryRes.data, taskOrder: updatedTaskOrder },
+            };
+          },
+        );
+        //add to new stack
+        queryClient.setQueryData(
+          ['stacks', over.id.toString()],
+          (stackQueryRes: AxiosResponse<Stack> | undefined) => {
+            if (!stackQueryRes) return;
+            const updatedTaskOrder = [...stackQueryRes.data.taskOrder];
+            updatedTaskOrder.push(active.id.toString());
+            return {
+              ...stackQueryRes,
+              data: { ...stackQueryRes.data, taskOrder: updatedTaskOrder },
+            };
+          },
+        );
+
+        setActiveTaskStackId(over.id.toString());
+      }
+    }
+  };
+
+  const onDragEnd = ({ active, over }: DragEndEvent) => {
+    console.log('dragEndActiveId :', active.id);
+    console.log('dragEndOverId :', over?.id);
+
+    setOverlayStackData(undefined);
+    setOverlayTaskData(undefined);
+
+    if (!over) return;
+
+    /* stackMove */
+    if (
+      active.data.current?.type === 'Stack' &&
+      over.data.current?.type === 'Stack' &&
+      active.id !== over.id
+    ) {
       queryClient.setQueryData(
         ['projects', projectId],
-        (projectQueryRes: AxiosResponse<Project>) => ({
-          ...projectQueryRes,
-          data: {
-            ...projectQueryRes.data,
-            stackOrder: stackOrder,
-          },
-        }),
-      );
-      return { prevQueryData };
-    },
-    onError: (_error, _data, context) => {
-      queryClient.setQueryData(['projects', projectId], context?.prevQueryData);
-    },
-  });
+        (projectQueryRes: AxiosResponse<Project>) => {
+          if (!projectQueryRes) return;
+          const stackOrder = projectQueryRes.data.stackOrder;
+          const activeIndex = stackOrder.findIndex(
+            (stackId) => stackId === active.id.toString(),
+          );
+          const overIndex = stackOrder.findIndex(
+            (stackId) => stackId === over.id.toString(),
+          );
+          const updatedStackOrder = arrayMove(
+            stackOrder,
+            activeIndex,
+            overIndex,
+          );
 
-  const taskMoveMutation = useMutation({
-    mutationFn: async ({
-      activeTaskId,
-      overTaskId,
-    }: {
-      activeTaskId: string;
-      overTaskId: string;
-    }) => {
-      await httpClient.post(
-        `/tasks/move?activeTaskId=${activeTaskId}&overTaskId=${overTaskId}`,
+          stackMoveMutation.mutate({
+            projectId: projectId,
+            stackOrder: updatedStackOrder,
+          });
+
+          return {
+            ...projectQueryRes,
+            data: { ...projectQueryRes.data, stackOrder: updatedStackOrder },
+          };
+        },
       );
-    },
-    onSuccess: () => {
-      console.log('taskMoveSuccess');
-      if (activeStackDataStart)
-        queryClient.invalidateQueries({
-          queryKey: ['stacks', activeStackDataStart.id],
-        });
-    },
-    onError: (error) => {
-      console.log('taskMoveError', error.message);
-    },
-  });
+
+      return;
+    }
+
+    /* taskMove */
+    if (active.data.current?.type === 'Task') {
+      queryClient.setQueryData(
+        ['tasks', active.id.toString()],
+        (taskQueryRes: AxiosResponse<Task> | undefined) => {
+          if (!taskQueryRes) return;
+          return {
+            ...taskQueryRes,
+            data: {
+              ...taskQueryRes.data,
+              stackId: activeTaskStackId,
+            },
+          };
+        },
+      );
+    }
+
+    if (
+      active.data.current?.type === 'Task' &&
+      over.data.current?.type === 'Task' &&
+      active.id !== over.id
+    ) {
+      //Update stack.stackOrder
+      queryClient.setQueryData(
+        ['stacks', over.data.current.taskData.stackId],
+        (stackQueryRes: AxiosResponse<Stack> | undefined) => {
+          if (!stackQueryRes) return;
+          const taskOrder = stackQueryRes.data.taskOrder;
+          const activeIndex = taskOrder.findIndex(
+            (taskId) => taskId === active.id.toString(),
+          );
+          const overIndex = taskOrder.findIndex(
+            (taskId) => taskId === over.id.toString(),
+          );
+          if (activeIndex === -1 || overIndex === -1) return;
+
+          const updatedTaskOrder = arrayMove(taskOrder, activeIndex, overIndex);
+          return {
+            ...stackQueryRes,
+            data: { ...stackQueryRes.data, taskOrder: updatedTaskOrder },
+          };
+        },
+      );
+
+      const toStackData = queryClient.getQueryData<
+        AxiosResponse<Stack> | undefined
+      >(['stacks', over.data.current.taskData.stackId])?.data;
+      if (!toStackData) return;
+
+      if (!fromTaskStackId) return;
+      const fromStackData = queryClient.getQueryData<
+        AxiosResponse<Stack> | undefined
+      >(['stacks', fromTaskStackId])?.data;
+      if (!fromStackData) return;
+
+      taskMoveMutation.mutate({
+        taskId: active.id.toString(),
+        toStackId: over.data.current.taskData.stackId,
+        toStackTaskOrder: toStackData.taskOrder,
+        fromStackId: fromTaskStackId,
+        fromStackTaskOrder: fromStackData.taskOrder,
+      });
+      return;
+    }
+
+    if (
+      active.data.current?.type === 'Task' &&
+      over.data.current?.type === 'Task' &&
+      active.id === over.id
+    ) {
+      if (!activeTaskStackId) return;
+      const toStackData = queryClient.getQueryData<
+        AxiosResponse<Stack> | undefined
+      >(['stacks', activeTaskStackId])?.data;
+      if (!toStackData) return;
+
+      if (!fromTaskStackId) return;
+      const fromStackData = queryClient.getQueryData<
+        AxiosResponse<Stack> | undefined
+      >(['stacks', fromTaskStackId])?.data;
+      if (!fromStackData) return;
+
+      taskMoveMutation.mutate({
+        taskId: active.id.toString(),
+        toStackId: activeTaskStackId,
+        toStackTaskOrder: toStackData.taskOrder,
+        fromStackId: fromTaskStackId,
+        fromStackTaskOrder: fromStackData.taskOrder,
+      });
+
+      return;
+    }
+
+    if (
+      active.data.current?.type === 'Task' &&
+      over.data.current?.type === 'Stack'
+    ) {
+      const toStackData = queryClient.getQueryData<
+        AxiosResponse<Stack> | undefined
+      >(['stacks', over.id.toString()])?.data;
+      if (!toStackData) return;
+
+      if (!fromTaskStackId) return;
+      const fromStackData = queryClient.getQueryData<
+        AxiosResponse<Stack> | undefined
+      >(['stacks', fromTaskStackId])?.data;
+      if (!fromStackData) return;
+
+      taskMoveMutation.mutate({
+        taskId: active.id.toString(),
+        toStackId: over.id.toString(),
+        toStackTaskOrder: toStackData.taskOrder,
+        fromStackId: fromTaskStackId,
+        fromStackTaskOrder: fromStackData.taskOrder,
+      });
+      return;
+    }
+  };
 
   const sensors = useSensors(
     useSensor(PointerSensor, {
@@ -104,196 +331,18 @@ export default function ProjectDndProvider({
       },
     }),
   );
-
-  const onDragStart = ({ active }: DragStartEvent) => {
-    if (active.data.current?.type === 'Stack') {
-      setActiveStackDataStart(active.data.current.stackData);
-    }
-    if (active.data.current?.type === 'Task') {
-      setActiveTaskDataStart(active.data.current.taskData);
-      setCurrentStackId(active.data.current.taskData.stackId);
-    }
-  };
-
-  const onDragOver = ({ active, over }: DragOverEvent) => {
-    if (!over) return;
-    if (active.id === over.id) return;
-    if (
-      active.data.current?.type === 'Task' &&
-      over.data.current?.type === 'Task'
-    ) {
-      const activeTaskData: Task =
-        queryClient.getQueryData<AxiosResponse<Task>>([
-          'tasks',
-          active.id.toString(),
-        ])?.data || active.data.current.taskData;
-
-      const overTaskData: Task = over.data.current.taskData;
-      setActiveTaskId(activeTaskData.id);
-      setOverTaskId(overTaskData.id);
-
-      if (currentStackId === overTaskData.stackId) {
-        queryClient.setQueryData(
-          ['stacks', overTaskData.stackId],
-          (stackQueryRes: AxiosResponse<Stack>) => {
-            const taskOrder = stackQueryRes.data.taskOrder;
-            const activeTaskIndex = taskOrder.findIndex(
-              (taskId) => taskId === activeTaskData.id,
-            );
-            const overTaskIndex = taskOrder.findIndex(
-              (taskId) => taskId === overTaskData.id,
-            );
-            return {
-              ...stackQueryRes,
-              data: {
-                ...stackQueryRes.data,
-                taskOrder: arrayMove(taskOrder, activeTaskIndex, overTaskIndex),
-              },
-            };
-          },
-        );
-
-        return;
-      }
-
-      //Delete task from current stack
-      queryClient.setQueryData(
-        ['stacks', currentStackId],
-        (stackQueryRes: AxiosResponse<Stack>) => {
-          const taskOrder = stackQueryRes.data.taskOrder.filter(
-            (taskId) => taskId !== activeTaskData.id,
-          );
-          return {
-            ...stackQueryRes,
-            data: { ...stackQueryRes.data, taskOrder: taskOrder },
-          };
-        },
-      );
-
-      //Add task to over stack
-      queryClient.setQueryData(
-        ['stacks', overTaskData.stackId],
-        (stackQueryRes: AxiosResponse<Stack>) => {
-          const taskOrder = [...stackQueryRes.data.taskOrder];
-          const overTaskIndex = taskOrder.findIndex(
-            (taskId) => taskId === overTaskData.id,
-          );
-          taskOrder.splice(
-            overTaskIndex,
-            1,
-            activeTaskData.id,
-            overTaskData.id,
-          );
-          return {
-            ...stackQueryRes,
-            data: { ...stackQueryRes.data, taskOrder: taskOrder },
-          };
-        },
-      );
-
-      //update task stack Id
-      queryClient.setQueryData(
-        ['tasks', overTaskData.id],
-        (taskQueryRes: AxiosResponse<Task>) => {
-          return (
-            taskQueryRes && {
-              ...taskQueryRes,
-              data: { ...taskQueryRes, stackId: overTaskData.stackId },
-            }
-          );
-        },
-      );
-
-      setCurrentStackId(overTaskData.stackId);
-    }
-  };
-
-  const onDragEnd = (event: DragEndEvent) => {
-    setActiveStackDataStart(undefined);
-    setActiveTaskDataStart(undefined);
-    setCurrentStackId(undefined);
-    const { active, over } = event;
-
-    console.log('activeTaskIdUseState', activeTaskId);
-    console.log('overTaskIdUseState', overTaskId);
-    if (!over) return;
-
-    //Dragging Stack
-    if (active.data.current?.type === 'Stack') {
-      const queryData = queryClient.getQueryData<AxiosResponse<Project>>([
-        'projects',
-        projectId,
-      ]);
-
-      const stackOrder = queryData?.data.stackOrder;
-      if (!stackOrder) return;
-
-      const activeIndex = stackOrder.findIndex(
-        (stackId) => stackId === active.id,
-      );
-      const overIndex = stackOrder.findIndex((stackId) => stackId === over.id);
-
-      stackOrderMutation.mutate({
-        projectId,
-        stackOrder: arrayMove(stackOrder, activeIndex, overIndex),
-      });
-    }
-
-    if (
-      active.data.current?.type === 'Task' &&
-      over.data.current?.type === 'Task'
-    ) {
-      if (activeTaskId && overTaskId) {
-        taskMoveMutation.mutate({
-          activeTaskId: activeTaskId,
-          overTaskId: overTaskId,
-        });
-      }
-    }
-
-    //Dragging Task
-    // if (
-    //   active.data.current?.type === 'Task' &&
-    //   over.data.current?.type === 'Task'
-    // ) {
-    //   const activeTaskData: Task = active.data.current.taskData;
-    //   const overTaskData: Task = over.data.current.taskData;
-
-    //   if (currentStackId === overTaskData.stackId) {
-    //     const stackQueryRes = queryClient.getQueryData<AxiosResponse<Stack>>([
-    //       'stacks',
-    //       activeTaskData.stackId,
-    //     ]);
-    //     const taskOrder = stackQueryRes.data.taskOrder;
-    //     if (!taskOrder) return;
-
-    //     const activeIndex = taskOrder.findIndex(
-    //       (taskId) => taskId === active.id,
-    //     );
-    //     const overIndex = taskOrder.findIndex((taskId) => taskId === over.id);
-
-    //     taskOrderMutation.mutate({
-    //       taskOrder: arrayMove(taskOrder, activeIndex, overIndex),
-    //       stackId: activeTaskData.stackId,
-    //     });
-    //   }
-    // }
-  };
-
   return (
     <DndContext
       sensors={sensors}
       onDragStart={onDragStart}
-      onDragOver={onDragOver}
       onDragEnd={onDragEnd}
+      onDragOver={onDragOver}
     >
       {children}
       {createPortal(
         <DragOverlay dropAnimation={null}>
-          {activeStackDataStart && (
-            <StackCard stackData={activeStackDataStart} />
-          )}
-          {activeTaskDataStart && <TaskCard taskData={activeTaskDataStart} />}
+          {overlayStackData && <StackCard stackData={overlayStackData} />}
+          {overlayTaskData && <TaskCard taskData={overlayTaskData} />}
         </DragOverlay>,
         document.body,
       )}
